星評価機能修正

+++ database/migrations/2023_12_25_043328_create_shops_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('address');
             $table->string('open_time');
             $table->text('menu');
-            $table->double('star', 8, 2);
+            $table->decimal('star', 3, 2);
             $table->timestamps();
         });
     }
